// src/interfaces/RutasResponse.tsx
// Generated by https://quicktype.io

export interface RutasResponse {
    rutas: Ruta[];
}

export interface Ruta {
    id:              string;
    numero:          number;
    destino:         string;
    origen:          string;
    estado:          boolean;
    longitudDestino: null | string;
    latitudDestino:  null | string;
    longitudOrigen:  null | string;
    latitudOrigen:   null | string;
}